Get rid of split unit_test_suite translation units; improve namespacing
