+++ test/buffer_manager_tests.cc
@@ -1,5 +1,4 @@
 #include "sycl_wrappers.h"
-#include "unit_test_suite_celerity.h"
 
 #include <catch2/catch_template_test_macros.hpp>
 #include <catch2/catch_test_macros.hpp>
@@ -47,7 +46,7 @@ namespace detail {
 		REQUIRE_FALSE(bm.has_buffer(b_id));
 
 		// TODO: check whether error was printed or not
-		maybe_print_graph(celerity::detail::runtime::get_instance().get_task_manager());
+		test_utils::maybe_print_graph(celerity::detail::runtime::get_instance().get_task_manager());
 	}
 
 // ComputeCPP based on Clang 8 segfaults in these tests
