+++ test/intrusive_graph_tests.cc
@@ -1,5 +1,3 @@
-#include "unit_test_suite_celerity.h"
-
 #include <catch2/catch_test_macros.hpp>
 
 #include "intrusive_graph.h"
