+++ test/system/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_executable(distr_tests distr_tests.cc)
 
-target_link_libraries(distr_tests PRIVATE unit_test_suite)
+target_link_libraries(distr_tests PRIVATE test_main)
 
 set_property(TARGET distr_tests PROPERTY CXX_STANDARD 17)
 set_property(TARGET distr_tests PROPERTY FOLDER "tests/system")
