+++ test/test_utils.h
@@ -356,6 +356,32 @@ namespace test_utils {
 		std::unique_ptr<detail::device_queue> dq;
 	};
 
+	// Printing of graphs can be enabled using the "--print-graphs" command line flag
+	inline bool print_graphs = false;
+
+	inline void maybe_print_graph(celerity::detail::task_manager& tm) {
+		if(print_graphs) {
+			const auto graph_str = tm.print_graph(std::numeric_limits<size_t>::max());
+			assert(graph_str.has_value());
+			CELERITY_INFO("Task graph:\n\n{}\n", *graph_str);
+		}
+	}
+
+	inline void maybe_print_graph(celerity::detail::command_graph& cdag) {
+		if(print_graphs) {
+			const auto graph_str = cdag.print_graph(std::numeric_limits<size_t>::max());
+			assert(graph_str.has_value());
+			CELERITY_INFO("Command graph:\n\n{}\n", *graph_str);
+		}
+	}
+
+	inline void maybe_print_graphs(celerity::test_utils::cdag_test_context& ctx) {
+		if(print_graphs) {
+			maybe_print_graph(ctx.get_task_manager());
+			maybe_print_graph(ctx.get_command_graph());
+		}
+	}
+
 } // namespace test_utils
 } // namespace celerity
